requirements.txt
feedparser==6.0.2
json5==0.9.5
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==3.2.1
blinker==1.4
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
gitdb==4.0.5
GitPython==3.1.12
idna==2.10
importlib-metadata==3.4.0
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.6
numpy==1.19.5
openpyxl==3.0.5
packaging==20.8
pandas==1.2.0
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.10
protobuf==3.14.0
ptyprocess==0.7.0
pyarrow==2.0.0
pycparser==2.20
pydeck==0.5.0
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.25.1
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.4
streamlit==0.74.1
terminado==0.9.2
testpath==0.4.4
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
validators==0.18.2
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==2.0.1
zipp==3.4.0
